+++ loggo/loggo.py
@@ -10,14 +10,7 @@ import uuid
 from contextlib import contextmanager
 from datetime import datetime
 from functools import wraps
-from typing import Optional, Dict, Union, Callable, Generator, Any, cast
-
-LOG_LEVELS = dict(critical='CRITICAL',
-                  dev='ERROR',
-                  error='ERROR',
-                  minor='WARNING',
-                  info='INFO',
-                  debug='DEBUG')
+from typing import Optional, Dict, Union, Callable, Generator, Any, cast, Mapping, Tuple
 
 # you don't need graylog installed
 try:
@@ -38,6 +31,8 @@ class Loggo:
     """
     # Callables with an attribute of this name set to True will not be logged by Loggo
     no_logs_attribute_name = '_do_not_log_this_callable'
+    # Only log when log level is this or higher
+    log_threshold = logging.DEBUG
 
     def __init__(self, config: Optional[Dict] = None) -> None:
         """
@@ -53,16 +48,16 @@ class Loggo:
         - truncation: truncate value of log data fields to this length
         - line_length: max length for console printed string
         - private_data: key names that should be filtered out of logging. when not
-        - max_dict_depth: how deep into log data loggo will look for private data
-        provided, nothing is censored
+        - max_dict_depth: how deep into log data loggo will look for private data provided, nothing is censored
         - raise_logging_errors: should Loggo errors be allowed to happen?
         - obscure: a string to use instead of any private data
+        - log_if_graylog_disabled: boolean value, should a warning log be made when failing to connect to graylog
         """
         config = config or dict()
         self.stopped = False
         self.allow_errors = True
         self.config = config
-        self.log_data = dict(loggo=True)
+        self.log_data = dict(loggo='True')
         self.facility = config.get('facility', 'loggo')
         self.ip = config.get('ip')
         self.port = config.get('port')
@@ -75,10 +70,10 @@ class Loggo:
         self.obscured = config.get('obscure', '[PRIVATE_DATA]')
         self.private_data = set(config.get('private_data', set()))
         self.max_dict_depth = config.get('max_dict_depth', 5)
-        self.no_graylog_disable_log = False
+        self.log_if_graylog_disabled = config.get('log_if_graylog_disabled', True)
         self.logger = logging.getLogger(self.facility)  # pylint: disable=no-member
-        self.logger.setLevel(logging.DEBUG)
-        self.add_handler()
+        self.logger.setLevel(Loggo.log_threshold)
+        self._add_graylog_handler()
 
     def _can_decorate(self, candidate: Callable, name: Optional[str] = None) -> bool:
         """
@@ -200,7 +195,7 @@ class Loggo:
         return self.logme(class_or_func, just_errors=True)
 
     def events(self, called: Optional[str] = None, returned: Optional[str] = None, errored: Optional[str] = None,
-               error_alert: str = 'dev') -> Callable:
+               error_level: int = logging.ERROR) -> Callable:
         """
         A decorator that takes messages as arguments
 
@@ -209,31 +204,31 @@ class Loggo:
         @Loggo.events(called='Log string for method call',
                       errored='Log string on exception',
                       returned='Log string for return',
-                      error_alert='critical')  # alert level for errors
+                      error_level=50)  # log level for errors
         def f():  # ...
         """
         def real_decorator(function: Callable) -> Callable:
             @wraps(function)
-            def wrapper(*args, **kwargs) -> Any:
+            def wrapper(*args: Any, **kwargs: Any) -> Any:
                 bound = self._params_to_dict(function, *args, **kwargs)
                 if bound is None:
                     return function(*args, **kwargs)
                 param_strings = self.sanitise(bound)
                 if called:
-                    self.log(called, None, param_strings)
+                    self.info(called, param_strings)
                 try:
                     ret = function(*args, **kwargs)
                     if returned:
                         ret_rep = self._represent_return_value(ret, truncate=500)
                         param_strings['return_value'] = ret_rep
                         param_strings['return_type'] = type(ret).__name__
-                        self.log(returned, None, param_strings)
+                        self.info(returned, param_strings)
                         return ret
                 except Exception as error:
                     if errored:
                         param_strings['error'] = str(error)
                         param_strings['trace'] = traceback.format_exc()
-                        self.log(errored, error_alert, param_strings)
+                        self.log(error_level, errored, param_strings)
             return wrapper
         return real_decorator
 
@@ -253,7 +248,7 @@ class Loggo:
             return function
 
         @wraps(function)
-        def full_decoration(*args, **kwargs) -> Any:
+        def full_decoration(*args: Any, **kwargs: Any) -> Any:
             """
             Main decorator logic. Generate a log before running the callable,
             then try to run it. If it errors, log the error. If it doesn't,
@@ -297,40 +292,38 @@ class Loggo:
                 raise
         return full_decoration
 
-    def _string_params(self, non_private_params, use_repr=True):
+    def _string_params(self, non_private_params: Dict, use_repr: bool = True) -> Dict[str, str]:
         """
         Turn every entry in log_data into truncated strings
         """
         params = dict()
         for key, val in non_private_params.items():
-            truncation = self.truncation if key not in {'trace', 'traceback'} else False
+            truncation = self.truncation if key not in {'trace', 'traceback'} else None
             safe_key = self._force_string_and_truncate(key, 50, use_repr=False)
             safe_val = self._force_string_and_truncate(val, truncation, use_repr=use_repr)
             params[safe_key] = safe_val
         return params
 
     @staticmethod
-    def _make_call_signature(function, param_strings):
+    def _make_call_signature(function: Callable, param_strings: Dict[str, str]) -> Tuple[str, Dict]:
         """
         Represent the call as a string mimicking how it is written in Python
         """
         signature = '{callable}({params})'
-        param_str = ', '.join('{}={}'.format(k, v) for k, v in param_strings.items())
+        param_str = ', '.join(f'{k}={v}' for k, v in param_strings.items())
         format_strings = dict(callable=getattr(function, '__qualname__', 'unknown_callable'),
                               params=param_str)
         formatted = signature.format(**format_strings)
         format_strings['call_signature'] = formatted
         return formatted, format_strings
 
-    def listen_to(loggo_self, facility, no_graylog_disable_log=False):
+    def listen_to(loggo_self, facility: str) -> None:
         """
         This method can hook the logger up to anything else that logs using the
         Python logging module (i.e. another logger) and steals its logs
         """
-        loggo_self.no_graylog_disable_log = no_graylog_disable_log
-
         class LoggoHandler(logging.Handler):
-            def emit(handler_self, record):
+            def emit(handler_self, record: logging.LogRecord) -> None:
                 attributes = {'msg', 'created', 'msecs', 'stack_info',
                               'levelname', 'filename', 'module', 'args',
                               'funcName', 'process', 'relativeCreated',
@@ -339,14 +332,13 @@ class Loggo:
                               'levelno'}
                 extra = dict(record.__dict__)
                 [extra.pop(attrib, None) for attrib in attributes]
-                alert = extra.get('alert')
                 extra['sublogger'] = facility
-                loggo_self.log(record.msg, alert, extra)
+                loggo_self.log(record.levelno, record.msg, extra)
         other_loggo = logging.getLogger(facility)
-        other_loggo.setLevel(logging.DEBUG)
+        other_loggo.setLevel(Loggo.log_threshold)
         other_loggo.addHandler(LoggoHandler())
 
-    def _params_to_dict(self, function, *args, **kwargs):
+    def _params_to_dict(self, function: Callable, *args: Any, **kwargs: Any) -> Mapping:
         """
         Turn args and kwargs into an OrderedDict of {param_name: value}
         """
@@ -360,7 +352,7 @@ class Loggo:
                 bound.pop('cls')
         return bound
 
-    def _obscure_private_keys(self, log_data, dict_depth=0):
+    def _obscure_private_keys(self, log_data: Any, dict_depth: int = 0) -> Any:
         """
         Obscure any private values in a dictionary recursively
         """
@@ -375,7 +367,7 @@ class Loggo:
                 out[key] = self._obscure_private_keys(value, dict_depth + 1)
         return out
 
-    def _represent_return_value(self, response, truncate=140):
+    def _represent_return_value(self, response: Any, truncate: Optional[int] = 140) -> str:
         """
         Make a string representation of whatever a method returns
         """
@@ -385,14 +377,14 @@ class Loggo:
 
         return '({})'.format(self._force_string_and_truncate(response, truncate, use_repr=True))
 
-    def _generate_log(self, where, returned, formatters, safe_log_data):
+    def _generate_log(self, where: str, returned: Any, formatters: Dict, safe_log_data: Dict[str, str]) -> None:
         """
         generate message, level and log data for automated logs
 
         where (str): 'pre'/'post'/'noreturn'/'error' --- the auto-log type
         returned (ANY): what the decorated callable returned
         formatters (dict): dict containing format strings needed for message
-        safe_log_data (bool): dict of stringified, truncated, censored parameters
+        safe_log_data (dict): dict of stringified, truncated, censored parameters
         """
         # if errors not to be shown and this is an error, quit
         if not self.allow_errors and where == 'error':
@@ -411,7 +403,7 @@ class Loggo:
 
         # return value for log message
         if where == 'post':
-            ret_str = self._represent_return_value(returned, truncate=False)
+            ret_str = self._represent_return_value(returned, truncate=None)
             formatters['return_value'] = ret_str
             formatters['return_type'] = type(returned).__name__
 
@@ -433,17 +425,18 @@ class Loggo:
         # turn it on just for now, as if we shouldn't log we'd have returned
         self.stopped = False
         # do logging
-        self.log(msg, extra=log_data, safe=True)
+        self.info(msg, extra=log_data, safe=True)
         # restore old stopped state
         self.stopped = original_state
 
-    def add_custom_log_data(self):
+    def add_custom_log_data(self) -> Dict[str, str]:
         """
         An overwritable method useful for adding custom log data
         """
         return dict()
 
-    def _build_string(self, msg, level, traceback=None):
+    @staticmethod
+    def _build_string(msg: str, level: int, trace: str = '') -> str:
         """
         Make a single line string, or multiline if traceback provided, for print
         and file logging
@@ -451,18 +444,18 @@ class Loggo:
         tstamp = datetime.now().strftime('%d.%m %Y %H:%M:%S')
         datapoints = [tstamp, msg, level]
         strung = '\t' + '\t'.join([str(s).strip('\n') for s in datapoints])
-        if traceback:
-            strung = '{} -- see below: \n{}\n'.format(strung, traceback)
+        if trace:
+            strung = f'{strung} -- see below: \n{trace}\n'
         return strung.strip('\n') + '\n'
 
-    def get_logfile(self, **kwargs):
+    def get_logfile(self, **kwargs: str) -> str:
         """
         This method exists so that it can be overwritten for applications requiring
         more complex logfile choices.
         """
         return self.logfile
 
-    def write_to_file(self, line, logfile=None):
+    def write_to_file(self, line: str, logfile: Optional[str] = None) -> None:
         """
         Very simple log writer, could expand. simple append the line to the file
         """
@@ -473,31 +466,31 @@ class Loggo:
         with open(logfile, 'a') as fo:
             fo.write(line.rstrip('\n') + '\n')
 
-    def add_handler(self):
-        """
-        Add a handler for Graylog
-        """
-        if not self.ip or not self.port or not graypy and not self.no_graylog_disable_log:
-            self.log('Graylog not configured! Disabling it')
+    def _add_graylog_handler(self) -> None:
+        if not self.ip or not self.port or not graypy:
+            if self.log_if_graylog_disabled:
+                self.warning('Graylog not configured! Disabling it')
             return
         handler = graypy.GELFUDPHandler(self.ip, self.port, debugging_fields=False)
         self.logger.addHandler(handler)
 
-    def _force_string_and_truncate(self, obj, truncate, use_repr=False):
+    def _force_string_and_truncate(self, obj: Any, truncate: Optional[int], use_repr: bool = False) -> str:
         """
-        Return stringified and truncated obj, or log alert if not possible
+        Return stringified and truncated obj. If stringification fails, log a warning
+        and return the string '<<Unstringable input>>'
         """
         try:
             obj = str(obj) if not use_repr else repr(obj)
         except Exception as error:
-            self.log('Object could not be cast to string', extra=dict(error_type=type(error), error=error))
+            self.warning('Object could not be cast to string', extra=dict(error_type=type(error), error=error))
             return '<<Unstringable input>>'
-        if truncate in {False, None}:
+        if truncate is None:
             return obj
         # truncate and return
         return (obj[:truncate] + '...') if len(obj) > (truncate + 3) else obj
 
-    def _rename_protected_keys(self, log_data):
+    @staticmethod
+    def _rename_protected_keys(log_data: Dict) -> Dict:
         """
         Some names cannot go into logger; remove them here and log the problem
         """
@@ -510,41 +503,30 @@ class Loggo:
             out[key] = value
         return out
 
-    def sanitise(self, unsafe_dict, use_repr=True):
+    def sanitise(self, unsafe_dict: Mapping, use_repr: bool = True) -> Dict[str, str]:
         """
         Ensure that log data is safe to log:
 
         - No private keys
         - Rename protected keys
-        - Everthing strings
+        - Everything strings
         """
         obscured = self._obscure_private_keys(unsafe_dict)
         no_protected = self._rename_protected_keys(obscured)
         return self._string_params(no_protected, use_repr=use_repr)
 
-    def sanitise_msg(self, msg):
+    def sanitise_msg(self, msg: str) -> str:
         """
         Overwritable method to clean or alter log messages
         """
         return msg
 
-    @staticmethod
-    def _get_log_level(alert):
-        """
-        Translate the passed in log level to an integer, default to 20
-        """
-        if isinstance(alert, str):
-            return getattr(logging, LOG_LEVELS.get(alert, 'INFO'))
-        elif isinstance(alert, int):
-            return alert
-        return 20
-
-    def log(self, message, alert=None, extra=None, safe=False):
+    def log(self, level: int, msg: str, extra: Optional[Dict] = None, safe: bool = False) -> None:
         """
         Main logging method, called both in auto logs and manually by user
 
-        message: string to log
-        alert: str/int priority of log
+        level: int, priority of log
+        msg: string to log
         extra: dict of extra fields to log
         safe: do we need to sanitise extra?
         """
@@ -556,20 +538,16 @@ class Loggo:
         extra = extra or dict()
         log_data = {**self.log_data, **extra}
 
-        # translate log levels to an integer --- things to fix here still
-        log_level = self._get_log_level(alert)
-
         if not safe:
             log_data = self.sanitise(log_data, use_repr=False)
-            message = self.sanitise_msg(message)
+            msg = self.sanitise_msg(msg)
 
-        log_data['alert'] = alert
+        log_data['log_level'] = str(level)
 
         # print or write log lines
-        line = None
         if self.do_print or self.do_write:
-            trace = extra.get('traceback')
-            line = self._build_string(message, alert, traceback=trace)
+            trace = extra.get('traceback', '')
+            line = self._build_string(msg, level, trace=trace)
         if self.do_print:
             print(line)
         if self.do_write:
@@ -580,8 +558,23 @@ class Loggo:
 
         # the only actual call to logging module's log method!
         try:
-            self.logger.log(log_level, message, extra=log_data)
+            self.logger.log(level, msg, extra=log_data)
         # it has been known to fail, e.g. when extra contains weird stuff
         except Exception:
             if self.raise_logging_errors:
                 raise
+
+    def debug(self, *args: Any, **kwargs: Any) -> None:
+        return self.log(logging.DEBUG, *args, **kwargs)
+
+    def info(self, *args: Any, **kwargs: Any) -> None:
+        return self.log(logging.INFO, *args, **kwargs)
+
+    def warning(self, *args: Any, **kwargs: Any) -> None:
+        return self.log(logging.WARNING, *args, **kwargs)
+
+    def error(self, *args: Any, **kwargs: Any) -> None:
+        return self.log(logging.ERROR, *args, **kwargs)
+
+    def critical(self, *args: Any, **kwargs: Any) -> None:
+        return self.log(logging.CRITICAL, *args, **kwargs)