+++ tests/tests.py
@@ -1,8 +1,9 @@
 import os
 import unittest
-from unittest.mock import mock_open, patch, ANY
+from unittest.mock import mock_open, patch, ANY, Mock
+import logging
 
-from loggo import Loggo as a_loggo
+from loggo import Loggo as LoggoType
 
 test_setup = dict(facility='LOGGO_TEST',
                   ip=None,
@@ -10,7 +11,7 @@ test_setup = dict(facility='LOGGO_TEST',
                   do_print=True,
                   do_write=True,
                   private_data=['mnemonic', 'priv'])
-Loggo = a_loggo(test_setup)
+Loggo = LoggoType(test_setup)
 
 
 @Loggo
@@ -71,7 +72,7 @@ all_method_types = AllMethodTypes()
 
 
 @Loggo
-class DummyClass(object):
+class DummyClass:
     """
     A class with regular methods, static methods and errors
     """
@@ -102,7 +103,7 @@ class DummyClass(object):
         raise ValueError('Bam!')
 
 
-class DummyClass2(object):
+class DummyClass2:
     def add(self, a, b, c):
         return a + b + c
 
@@ -159,7 +160,7 @@ class TestDecoration(unittest.TestCase):
             with self.assertRaises(ValueError):
                 first_test_func(5)
             (alert, logged_msg), extras = logger.call_args_list[-1]
-            self.assertEqual(logged_msg, '*Errored during first_test_func(number=5) with ValueError "Broken!"', logged_msg)
+            self.assertEqual(logged_msg, '*Errored during first_test_func(number=5) with ValueError "Broken!"')
 
     def test_log_errors(self):
         with patch('logging.Logger.log'):
@@ -176,7 +177,9 @@ class TestDecoration(unittest.TestCase):
                 may_or_may_not_error_test('astadh', 1331)
             (alert, logged_msg), extras = logger.call_args
             self.assertEqual(alert, 20)
-            self.assertEqual(logged_msg, '*Errored during may_or_may_not_error_test(first=\'astadh\', other=1331) with ValueError "no good"', logged_msg)
+            expected_msg = ('*Errored during may_or_may_not_error_test(first=\'astadh\', other=1331) '
+                            'with ValueError "no good"')
+            self.assertEqual(logged_msg, expected_msg)
 
     def test_logme_0(self):
         """
@@ -189,7 +192,9 @@ class TestDecoration(unittest.TestCase):
             (alert, logged_msg), extras = logger.call_args_list[0]
             self.assertEqual(logged_msg, '*Called may_or_may_not_error_test(first=2534, other=2466, kwargs=True)')
             (alert, logged_msg), extras = logger.call_args_list[-1]
-            self.assertEqual(logged_msg, '*Returned from may_or_may_not_error_test(first=2534, other=2466, kwargs=True) with tuple ((5000, True))', logged_msg)
+            expected_msg = ('*Returned from may_or_may_not_error_test(first=2534, other=2466, '
+                            'kwargs=True) with tuple ((5000, True))')
+            self.assertEqual(logged_msg, expected_msg)
 
     def test_logme_1(self):
         with patch('logging.Logger.log') as logger:
@@ -198,7 +203,7 @@ class TestDecoration(unittest.TestCase):
             (alert, logged_msg), extras = logger.call_args_list[0]
             self.assertEqual(logged_msg, '*Called DummyClass.add(a=1, b=2)')
             (alert, logged_msg), extras = logger.call_args_list[-1]
-            self.assertEqual('*Returned from DummyClass.add(a=1, b=2) with int (3)', logged_msg, logged_msg)
+            self.assertEqual('*Returned from DummyClass.add(a=1, b=2) with int (3)', logged_msg)
 
     def test_everything_0(self):
         with patch('logging.Logger.log') as logger:
@@ -206,14 +211,15 @@ class TestDecoration(unittest.TestCase):
             (alert, logged_msg), extras = logger.call_args_list[0]
             self.assertEqual(logged_msg, '*Called DummyClass.add_and_maybe_subtract(a=15, b=10, c=5)')
             (alert, logged_msg), extras = logger.call_args_list[-1]
-            self.assertEqual('*Returned from DummyClass.add_and_maybe_subtract(a=15, b=10, c=5) with int (20)', logged_msg, logged_msg)
+            expected_msg = '*Returned from DummyClass.add_and_maybe_subtract(a=15, b=10, c=5) with int (20)'
+            self.assertEqual(expected_msg, logged_msg)
 
     def test_everything_1(self):
         with patch('logging.Logger.log') as logger:
             result = dummy.static_method(10)
             self.assertEqual(result, 100)
             (alert, logged_msg), extras = logger.call_args_list[-1]
-            self.assertEqual('*Returned from DummyClass.static_method(number=10) with int (100)', logged_msg, logged_msg)
+            self.assertEqual('*Returned from DummyClass.static_method(number=10) with int (100)', logged_msg)
 
     def test_everything_3(self):
         with patch('logging.Logger.log') as logger:
@@ -274,7 +280,7 @@ class TestDecoration(unittest.TestCase):
             self.assertFalse('secret' in extras['extra']['data'])
 
 
-class NoRepr(object):
+class NoRepr:
     """
     An object that really hates being repr'd
     """
@@ -285,8 +291,11 @@ class NoRepr(object):
 class TestLog(unittest.TestCase):
 
     def setUp(self):
+        self.log_msg = 'This is a message that can be used when the content does not matter.'
+        self.log_data = {'This is': 'log data', 'that can be': 'used when the content does not matter'}
+
         self.test_setup = dict(facility='LOG_TEST', ip=None, port=None, do_print=True, do_write=True)
-        self.loggo = a_loggo(self.test_setup)
+        self.loggo = LoggoType(self.test_setup)
         self.log = self.loggo.log
 
     def test_protected_keys(self):
@@ -295,22 +304,21 @@ class TestLog(unittest.TestCase):
         in order to stop error in logger later
         """
         with patch('logging.Logger.log') as mock_log:
-            self.log('fine', None, dict(name='bad', other='good'))
+            self.log(logging.INFO, 'fine', dict(name='bad', other='good'))
             (_alert, _msg), kwargs = mock_log.call_args
             self.assertEqual(kwargs['extra']['protected_name'], 'bad')
             self.assertEqual(kwargs['extra']['other'], 'good')
 
     def test_can_log(self):
         with patch('logging.Logger.log') as logger:
-            nums = [(None, 20), ('dev', 40), ('critical', 50)]
+            level_num = 50
             msg = 'Test message here'
-            for level, num in nums:
-                result = self.log(msg, level, dict(extra='data'))
-                self.assertIsNone(result)
-                (alert, logged_msg), extras = logger.call_args
-                self.assertEqual(alert, num)
-                self.assertEqual(msg, logged_msg)
-                self.assertEqual(extras['extra']['extra'], 'data')
+            result = self.log(level_num, msg, dict(extra='data'))
+            self.assertIsNone(result)
+            (alert, logged_msg), extras = logger.call_args
+            self.assertEqual(alert, level_num)
+            self.assertEqual(msg, logged_msg)
+            self.assertEqual(extras['extra']['extra'], 'data')
 
     def test_write_to_file(self):
         """
@@ -318,7 +326,7 @@ class TestLog(unittest.TestCase):
         """
         mock = mock_open()
         with patch('builtins.open', mock):
-            self.log('An entry in our log')
+            self.log(logging.INFO, 'An entry in our log')
             mock.assert_called_with(Loggo.logfile, 'a')
             self.assertTrue(os.path.isfile(Loggo.logfile))
 
@@ -330,7 +338,7 @@ class TestLog(unittest.TestCase):
             msg = 'This is simply a test of the int truncation inside the log.'
             large_number = 10**300001
             log_data = dict(key=large_number)
-            self.log(msg, None, log_data)
+            self.log(logging.INFO, msg, log_data)
             mock_log.assert_called_with(20, msg, extra=ANY)
             logger_was_passed = mock_log.call_args[1]['extra']['key']
             # 7500 here is the default self.truncation for loggo
@@ -363,14 +371,14 @@ class TestLog(unittest.TestCase):
             mock_log.side_effect = Exception('Really dead.')
             self.loggo.raise_logging_errors = True
             with self.assertRaises(Exception):
-                self.loggo.log('Anything')
+                self.loggo.log(logging.INFO, 'Anything')
 
     def test_loggo_pause(self):
         with patch('logging.Logger.log') as mock_log:
             with Loggo.pause():
-                Loggo.log('test')
+                Loggo.log(logging.INFO, 'test')
             mock_log.assert_not_called()
-            Loggo.log('test')
+            Loggo.log(logging.INFO, 'test')
             mock_log.assert_called()
 
     def test_loggo_pause_error(self):
@@ -379,7 +387,9 @@ class TestLog(unittest.TestCase):
                 with self.assertRaises(ValueError):
                     may_or_may_not_error_test('one', 'two')
             (alert, msg), kwargs = logger.call_args
-            self.assertEqual('*Errored during may_or_may_not_error_test(first=\'one\', other=\'two\') with ValueError "no good"', msg, msg)
+            expected_msg = ('*Errored during may_or_may_not_error_test(first=\'one\', '
+                            'other=\'two\') with ValueError "no good"')
+            self.assertEqual(expected_msg, msg)
             logger.assert_called_once()
             logger.reset()
             with self.assertRaises(ValueError):
@@ -392,24 +402,24 @@ class TestLog(unittest.TestCase):
                 with self.assertRaises(ValueError):
                     may_or_may_not_error_test('one', 'two')
             logger.assert_not_called()
-            Loggo.log('test')
+            Loggo.log(logging.INFO, 'test')
             logger.assert_called_once()
 
     def test_see_below(self):
         msg = 'testing only'
-        s = self.loggo._build_string(msg, 'dev', traceback=False)
+        s = self.loggo._build_string(msg, 50)
         self.assertTrue('-- see below:' not in s)
 
     def test_compat(self):
         test = 'a string'
         with patch('loggo.Loggo.log') as logger:
-            Loggo.log(test, None, None)
+            Loggo.log(logging.INFO, test, None)
         args = logger.call_args
-        self.assertEqual(args[0][0], test)
-        self.assertIsNone(args[0][1])
+        self.assertIsInstance(args[0][0], int)
+        self.assertEqual(args[0][1], test)
         self.assertIsNone(args[0][2])
         with patch('logging.Logger.log') as logger:
-            Loggo.log(test)
+            Loggo.log(logging.INFO, test)
         (alert, msg), kwargs = logger.call_args
         self.assertEqual(test, msg)
 
@@ -420,6 +430,42 @@ class TestLog(unittest.TestCase):
             return needed
         with self.assertRaises(TypeError):
             dummy()
+
+    def test_debug(self):
+        with patch('loggo.Loggo.log') as logger:
+            self.loggo.debug(self.log_msg, self.log_data)
+            logger.assert_called_with(logging.DEBUG, self.log_msg, self.log_data)
+
+    def test_info(self):
+        with patch('loggo.Loggo.log') as logger:
+            self.loggo.info(self.log_msg, self.log_data)
+            logger.assert_called_with(logging.INFO, self.log_msg, self.log_data)
+
+    def test_warning(self):
+        with patch('loggo.Loggo.log') as logger:
+            self.loggo.warning(self.log_msg, self.log_data)
+            logger.assert_called_with(logging.WARNING, self.log_msg, self.log_data)
+
+    def test_error(self):
+        with patch('loggo.Loggo.log') as logger:
+            self.loggo.error(self.log_msg, self.log_data)
+            logger.assert_called_with(logging.ERROR, self.log_msg, self.log_data)
+
+    def test_critical(self):
+        with patch('loggo.Loggo.log') as logger:
+            self.loggo.critical(self.log_msg, self.log_data)
+            logger.assert_called_with(logging.CRITICAL, self.log_msg, self.log_data)
+
+    def test_listen_to(self):
+        sub_loggo_facility = 'a sub logger'
+        sub_loggo = LoggoType({'facility': sub_loggo_facility})
+        self.loggo.listen_to(sub_loggo_facility)
+
+        self.loggo.log = Mock()
+        log_args = logging.WARNING, 'The parent logger should log this message after sublogger logs it'
+        sub_loggo.log(*log_args)
+
+        self.loggo.log.assert_called_with(*log_args, ANY)
 
 
 class TestMethods(unittest.TestCase):